% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_add_data.R
\name{mod_add_data_ui}
\alias{mod_add_data_ui}
\alias{mod_add_data_server}
\title{mod_add_data_ui and mod_add_data_server}
\usage{
mod_add_data_ui(id, label = "Add Occurrence Data")

mod_add_data_server(input, output, session,
  next_button_id = "dataToConfigureDiv")
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
