% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_citation.R
\name{mod_citation_ui}
\alias{mod_citation_ui}
\alias{mod_citation_server}
\title{mod_citation_ui and mod_citation_server}
\usage{
mod_citation_ui(id, package)

mod_citation_server(input, output, session, package)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
