% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_darwinize.R
\name{mod_darwinize_ui}
\alias{mod_darwinize_ui}
\alias{mod_darwinize_server}
\title{module to add darwinization ability}
\usage{
mod_darwinize_ui(id)

mod_darwinize_server(input, output, session, dat)
}
\arguments{
\item{id}{shiny id}
}
\description{
UI to add darwinization funtion.
}
\keyword{internal}
